chore(generate-examples): generate fixtures (#17737)
Motivation: we want the ability for `rosetta:extract` to be called after `generate-examples`. Currently, this results in errors because the generated examples are only stored in the assembly, and only the visible source, so they fail compilation when we attempt to extract them.

Solution: while we generate examples, we also generate a fixture per assembly, `_generated.ts-fixture`, that stores the necessary import statement for compilation. Then, when we later (might) call extract, we have a fixture to place the example in.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*